[ZEPPELIN-2410]: Using UUID from filename in registry cache dir instead of uri.
The previous version of this file used the full uri as file name in the
cache. This included special characters like the : in http: which
caused problems.

Since this is just a cache file, a better approach is to use an UUID.

The UUID is constructed from the URI to avoid the cache dir to grow
over time (in case there is no clean up of the cache dir).