[ZEPPELIN-2438]: Add local winutils.exe in build and use if required on run.
The build of the spark dependencies now downloads winutils.exe and
installs it in interpreter/spark/winutils.
The zeppelin-env.cmd (only used if we run on windows) then checks if
HADOOP_HOME exists, if not, it will set it to point to winutils. This
is required to have the spark interpreter work in Windows.

In order to have this active, the zeppelin-env.cmd had to be removed
from the .gitignore.